+++ MonitorizareVot/MVNote.swift
@@ -5,7 +5,7 @@ import UIKit
 
 class MVNote {
     var sectionInfo: MVSectionInfo
-    var questionID: Int?
+    var questionID: Int16?
     var body: String?
     var image: UIImage?
     var synced: Bool = false
